+++ scripts/updateLinksCatalog.py
@@ -5,7 +5,6 @@ from model import Link
 from main import loadEnvFile
 from managers import DBManager
 
-
 def main():
     '''
     Updating NYPL Link media type with new link to replace text/html link
@@ -26,12 +25,10 @@ def main():
     dbManager.createSession()
 
     for link in dbManager.session.query(Link) \
-        .filter(Link.media_type == 'text/html').all():
-            
-            if link.flags != None and len(link.flags) > 0:
-                for i in link.flags:
-                    if i['catalog'] == True:
-                        link.media_type = 'application/html/catalog'
+        .filter(Link.media_type == 'text/html') \
+        .filter(Link.flags != {}) \
+        .filter(Link.flags['catalog'] == 'true').all():
+            link.media_type = 'application/html+catalog'
                 
 
     dbManager.commitChanges()