+++ scripts/updateRecordsCatalogLink.py
@@ -25,19 +25,19 @@ def main():
 
     dbManager.createSession()
 
-    catalogRegex = r'catalog: true'
-    linkRegex = r'text/html'
+    catalogRegex = r'catalog\": true' 
+    linkRegex = r'text\/html'
 
     for record in dbManager.session.query(Record) \
         .filter(Record.source == 'nypl').all():
             
-            for i in record.has_part:
-                if re.search(catalogRegex, i) != None:
-                    if re.search(linkRegex, i) != None:
-                        i.replace('text/html', 'application/html/catalog')
-                
+            for i, elem in enumerate(record.has_part):
+                if re.search(linkRegex, elem) != None:
+                    if re.search(catalogRegex, elem) != None:
+                        record.has_part = [record.has_part[i].replace('text/html', 'application/html+catalog')]
 
     dbManager.commitChanges()
+    
 
 if __name__ == '__main__':
     main()